+++ packages/test-utils/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @pie-qti/test-utils
 
+## 0.1.20
+
+### Patch Changes
+
+- @pie-qti/logger@0.1.20
+- @pie-qti/storage@0.1.20
+- @pie-qti/transform-core@0.1.20
+- @pie-qti/transform-types@0.1.20
+
 ## 0.1.19
 
 ### Patch Changes