+++ tools/cli/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @pie-qti/transform-cli
 
+## 0.1.20
+
+### Patch Changes
+
+- @pie-qti/ims-cp-node@0.1.20
+- @pie-qti/logger@0.1.20
+- @pie-qti/to-pie@0.1.20
+- @pie-qti/transform-core@0.1.20
+- @pie-qti/transform-types@0.1.20
+
 ## 0.1.19
 
 ### Patch Changes